Improve Dialogs & Refresh Tags & Things

+++ angular/src/app/things/things-controller.js
@@ -17,7 +17,7 @@ angular
 	$controller('TagsBaseCtrl', { $scope: $scope });
 
 	$scope.showThingDetailView = function showThingsDetailView(id) {
-    console.log("$scope.showThingDetailView")
+		console.log("$scope.showThingDetailView")
 		$location.path('/things/'+ id + '/view');
 	};
 
@@ -68,32 +68,23 @@ angular
 	$scope.thing = $scope.thing || {};
 	$scope.tags = $scope.tags || [];
 
-  $scope.searchTags = function searchTags(keywords) {
-  	keywords = keywords ? keywords.split(' ') : [];
-  	return api.searchTags(keywords);
-  }
-
-
-    $scope.dialogShown = false;
-
-	function showDetailsDialog(thing) {
-		if (!$scope.dialogShown) {
-			var dialog = ngDialog.open({
-				templateUrl: '/utag/things/thing-edit-directive.html',
-				controller: 'ThingsUpdateCtrl',
-				className: 'ngdialog-theme-plain',
-				scope: $scope,
-			});
-			$scope.dialogShown = true;
-
-			dialog.closePromise.then(function (data) {
-				$scope.dialogShown = false;
-				// $log.info(data.id + ' has been dismissed.');
-			});
-		}
+	$scope.searchTags = function searchTags(keywords) {
+		keywords = keywords ? keywords.split(' ') : [];
+		return api.searchTags(keywords);
 	}
 
+	$scope.showCreateDialog = showCreateDialog;
 	$scope.showDetailsDialog = showDetailsDialog;
+	$scope.showDeleteDialog = showDeleteDialog;
+
+	$scope.saveThing = function saveThing(thing) {
+		Things.repo.update({id: thing.id}, thing, function(data) {$location.path('/');}, function(data) {
+			$scope.messages = data.data.errors;
+			console.log("failAtUpdate");
+		});
+	};
+
+	$scope.dialogShown = false;
 
 	function showCreateDialog() {
 		if (!$scope.dialogShown) {
@@ -111,29 +102,71 @@ angular
 		}
 	}
 
-	$scope.showCreateDialog = showCreateDialog;
+	function showDetailsDialog(thing) {
+		if (!$scope.dialogShown) {
+			var dialog = ngDialog.open({
+				name: 'detail',
+				templateUrl: '/utag/things/thing-edit-directive.html',
+				controller: 'ThingsUpdateCtrl',
+				className: 'ngdialog-theme-plain',
+				scope: $scope,
+			});
+			$scope.dialogShown = true;
 
-  $scope.saveThing = function saveThing(thing) {
-		Things.repo.update({id: thing.id}, thing, function(data) {$location.path('/');}, function(data) {
-      $scope.messages = data.data.errors;
-      console.log("failAtUpdate");
-    });
-	};
+			dialog.closePromise.then(function (data) {
+				$scope.dialogShown = false;
+				// refresh tags
+				$scope.$parent.fetchTags();
+				$log.info(data.id + ' has been dismissed.');
+			});
+		}
+	}
 
-  //$scope.dialogShown = false;
-  $scope.showDeleteDialog = function showDeleteDialog(thing) {
-    Things.repo.delete({id: thing.id}, thing, function(data) {$location.path('/');}, function(data) {
-      $scope.messages = data.data.errors;
-      console.log("failAtDelete");
-    });
-  }
+	function showDeleteDialog(thing) {
+		ngDialog.openConfirm({
+			name: 'delete',
+			controller: 'ThingsUpdateCtrl',
+			className: 'ngdialog-theme-plain',
+			template:'\
+				<p>Are you sure you want to delete the Thing "{{thing.name}}" with all its tags?</p>\
+				<div class="ngdialog-buttons">\
+						<button type="button" class="ngdialog-button ngdialog-button-secondary" data-ng-click="closeThisDialog(0)">No</button>\
+						<button type="button" class="ngdialog-button ngdialog-button-primary" data-ng-click="confirm(1)">Yes</button>\
+				</div>',
+			plain: true,
+			scope: $scope,
+		}).then(function (value) {
+			$log.info('Modal promise resolved. Value: ', value);
+			Things.repo.delete({id: thing.id}, thing, function(data) {$location.path('/');}, function(data) {
+				$scope.messages = data.data.errors;
+				$log.info("failAtDelete");
+			});
 
-  activate();
+			// remove deleted things
+			for (var i = $scope.filteredThings.length - 1; i >= 0; i--) {
+				if ($scope.filteredThings[i].id === thing.id) {
+					$scope.filteredThings.splice(i, 1);
+					break;
+				}
+			};
+
+			// refresh tags
+			$scope.$parent.fetchTags();
+
+			ngDialog.close('detail');
+		}, function (reason) {
+			$log.info('Modal promise rejected. Reason: ', reason);
+		});
+
+		return true;
+	}
+
+	activate();
 
 	function activate() {
 		if ($routeParams.id) {
 			Things.repo.get({id: $routeParams.id}, function(data) {
-					$scope.thing = data;
+				$scope.thing = data;
 			});
 		}
 		if ($scope.tags.length === 0) {
@@ -145,115 +178,80 @@ angular
 
 /**
  * @ngdoc function
- * @name utag.tags.controller:ThingsDetailCtrl
+ * @name utag.tags.controller:ThingsUpdateCtrl
  * @description
- * # ThingsDetailCtrl
+ * # ThingsUpdateCtrl
  * Controller of the utag app
  */
 .controller('ThingsUpdateCtrl', function ThingsUpdateCtrl ($scope, $log, $controller, $routeParams, $location, ngDialog, api, Tags, Things) {
-  'use strict';
+	'use strict';
 
-  // extend ThingsCtrl
-  $controller('ThingsBaseCtrl', { $scope: $scope });
+	// extend ThingsCtrl
+	$controller('ThingsBaseCtrl', { $scope: $scope });
 
-  $scope.title = 'ThingDetail';
-  $scope.thing = $scope.thing || {};
-  $scope.tags = $scope.tags || [];
+	$scope.title = 'ThingUpdate';
+	$scope.thing = $scope.thing || {};
+	$scope.tags = $scope.tags || [];
 
-  $scope.searchTags = function searchTags(keywords) {
-    keywords = keywords ? keywords.split(' ') : [];
-    return api.searchTags(keywords);
-  }
+	$scope.searchTags = function searchTags(keywords) {
+		keywords = keywords ? keywords.split(' ') : [];
+		return api.searchTags(keywords);
+	}
 
-  $scope.saveThing = function saveThing(thing) {
-    Things.repo.update({id: $scope.thing.id}, thing, function(data) {$location.path('/');}, function(data) {
-      $scope.messages = data.data.errors;
-      console.log("failAtUpdate");
-    });
-  };
+	$scope.saveThing = function saveThing(thing) {
+		Things.repo.update({id: $scope.thing.id}, thing, function(data) {
+			$scope.closeThisDialog();
+			$location.path('/');
+		}, function(data) {
+			$scope.messages = data.data.errors;
+			console.log("failAtUpdate");
+		});
+	};
 
-  //$scope.dialogShown = false;
-  $scope.showDeleteDialog = function showDeleteDialog(thing) {
-    Things.repo.delete({id: $scope.thing.id}, thing, function(data) {$location.path('/');}, function(data) {
-      $scope.messages = data.data.errors;
-      console.log("failAtDelete");
-    });
+	activate();
 
-   /* if (!$scope.dialogShown) {
-      var dialog = ngDialog.open({
-        templateUrl: '/utag/things/thing-delete.html',
-        controller: 'ThingsDetailCtrl',
-        className: 'ngdialog-theme-plain',
-        scope: $scope,
-      });
-      $scope.dialogShown = true;
-
-      dialog.closePromise.then(function (data) {
-        $scope.dialogShown = false;
-        if (data == '1') {
-          Things.repo.update({id: $routeParams.id}, thing, function(data) {$location.path('/');}, function(data) {console.log("fail")});
-        }
-        // $log.info(data.id + ' has been dismissed.');
-      });
-    }*/
-
-    $scope.yesDelete = function yesDelete(thing) {
-      Things.repo.update({id: $scope.thing.id}, thing, function(data) {$location.path('/');}, function(data) {
-        $scope.messages = data.data.errors;
-        console.log("failAtyesDelete");
-      });
-    }
-
-    $scope.noDelete = function noDelete() {
-      $scope.dialogShown = false;
-    }
-
-  }
-
-  activate();
-
-  function activate() {
-    if ($scope.tags.length === 0) {
-      $scope.tags = api.userTagsDistinct();
-    }
-  }
+	function activate() {
+		if ($scope.tags.length === 0) {
+			$scope.tags = api.userTagsDistinct();
+		}
+	}
 
 })
 
 
-.controller('ThingsCreateCtrl', function ThingsDetailCtrl ($scope, $log, $controller, $routeParams, $location, ngDialog, api, Tags, Things) {
-  'use strict';
+.controller('ThingsCreateCtrl', function ThingsCreateCtrl ($scope, $log, $controller, $routeParams, $location, ngDialog, api, Tags, Things) {
+	'use strict';
 
-  // extend ThingsCtrl
-  $controller('ThingsBaseCtrl', { $scope: $scope });
+	// extend ThingsCtrl
+	$controller('ThingsBaseCtrl', { $scope: $scope });
 
-  $scope.title = 'ThingDetail';
-  $scope.thing = {name: '', description: '', tags: [], thingable: {id: '', uri: ''}};
-  $scope.tags = $scope.tags || [];
+	$scope.title = 'ThingDetail';
+	$scope.thing = {name: '', description: '', tags: [], thingable: {id: '', uri: ''}};
+	$scope.tags = $scope.tags || [];
 
 
-  $scope.searchTags = function searchTags(keywords) {
-    keywords = keywords ? keywords.split(' ') : [];
-    return api.searchTags(keywords);
-  }
+	$scope.searchTags = function searchTags(keywords) {
+		keywords = keywords ? keywords.split(' ') : [];
+		return api.searchTags(keywords);
+	}
 
-  $scope.createThing = function createThing(thing) {
-    Things.repo.save('', thing, function(data) {$location.path('/');}, function(data) {
-      $scope.messages = data.data.errors;
-    });
-  };
+	$scope.createThing = function createThing(thing) {
+		Things.repo.save('', thing, function(data) {$location.path('/');}, function(data) {
+			$scope.messages = data.data.errors;
+		});
+	};
 
-  $scope.cancel = function cancel() {
-    $location.path('/');
-  };
+	$scope.cancel = function cancel() {
+		$location.path('/');
+	};
 
-  activate();
+	activate();
 
-  function activate() {
-    if ($scope.tags.length === 0) {
-      $scope.tags = api.userTagsDistinct();
-    }
-  }
+	function activate() {
+		if ($scope.tags.length === 0) {
+			$scope.tags = api.userTagsDistinct();
+		}
+	}
 
 });
